requirements.txt
aiosqlite~=0.21.0
pydantic~=2.11.7
discord.py~=2.5.2
python-dotenv~=1.1.1